requirements.txt
google-generativeai>=0.8.0
langgraph>=0.2.0
langchain-core>=0.3.0
pydantic>=2.7.0
chromadb>=0.5.0


